+++ package.json
@@ -11,10 +11,17 @@
     "react": "^15.6.1",
     "react-addons-css-transition-group": "^15.6.0",
     "react-dom": "^15.6.1",
+    "react-router-dom": "^4.1.1",
     "webpack": "^3.0.0",
     "webpack-dev-server": "^2.5.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "css-loader": "^0.28.4",
+    "jsx-loader": "^0.13.2",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.4",
+    "style-loader": "^0.18.2"
+  },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "webpack -d --watch",
