+++ src/client/app/Components/Button/style.less
@@ -1,0 +1,3 @@
+button {
+    color: blue;
+}
