+++ src/client/app/Components/List/index.jsx
@@ -3,14 +3,14 @@ import Item from '../Item/index.jsx'
 
 class List extends Component {
     render () {
-        let itemArray = this.props.listItem.name.map(item => {
+        let listArray = this.props.listItem.name.map(item => {
             console.log("item",item);
             return ( <Item name={ item } />)
         })
         return (
             <div className="listWrapper">
                 <ul>
-                    {itemArray}
+                    {listArray}
                 </ul>
             </div>
         );
