+++ src/client/app/Pages/LandingPage/styles.less
@@ -1,16 +1,3 @@
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Friendly Rolls</title>
-  </head>
-  <body>
-    <div id="app" />
-    <script src="client/public/bundle.js" type="text/javascript"></script>
-  </body>
-
-
-<style>
-
 .modal-anim-enter {
     opacity: 0.00;
     transform: scale(0.7);
@@ -46,7 +33,3 @@
   width: 100%;
   height: 100%;
 }
-</style>
-</html>
-
-
