+++ src/client/public/index.html
@@ -1,0 +1,12 @@
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Friendly Rolls</title>
+  </head>
+  <body>
+    <div id="app" />
+    <script src="./bundle.js" type="text/javascript"></script>
+  </body>
+</html>
+
+
