+++ webpack.config.js
@@ -16,12 +16,25 @@ var config = {
         test : /\.jsx?/,
         include : APP_DIR,
         loader : 'Babel-loader'
+      },
+      { test: /\.less$/, 
+        use: [{
+          loader: "style-loader" // creates style nodes from JS strings
+        }, {
+          loader: "css-loader" // translates CSS into CommonJS
+        }, {
+          loader: "less-loader" // compiles Less to CSS
+        }]
+ 
       }
     ]
   },
   devServer: {
+    contentBase: path.join(__dirname, "src/client/public"),
     historyApiFallback: true,
+
   },
+  //Add a resolve path here to clean up the import and exportss
 };
 
 module.exports = config;
